restructure code into t_lexer module
